requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
pydantic==2.12.0
httpx>=0.27
requests==2.32.5
openai==1.109.1
slowapi==0.1.9
starlette==0.52.1
limits==5.8.0
redis>=5.0